requirements.txt
humanize ~= 4.3.0
im-squonk2-client >= 1.7.0, < 2.0.0
pandas ~= 1.4.0
pyyaml ~= 6.0
textual == 0.1.18
